+++ src/OnboardingSPA/components/SiteGenError/stylesheet.scss
@@ -137,5 +137,12 @@
 			}
 		}
 
+
+		&__migrationerror {
+
+			.nfd-onboarding-step--site-gen__error__container {
+				width: 610px;
+			}
+		}
 	}
 }
